ajustes nos eventos

+++ src/Prof.Hub.Domain/Aggregates/Teacher/Events/TeacherRatingUpdatedEvent.cs
@@ -1,5 +1,5 @@
 ﻿using Prof.Hub.SharedKernel;
 
 namespace Prof.Hub.Domain.Aggregates.Teacher.Events;
-public record TeacherRatingUpdatedEvent(string TeacherId, int NewRating) : IDomainEvent;
+public record TeacherRatingUpdatedEvent(string TeacherId, decimal NewRating) : IDomainEvent;
